Refactored the test class
