+++ hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/services/AbfsClient.java
@@ -1220,7 +1220,7 @@ public class AbfsClient implements Closeable {
    * Getter for abfsCounters from AbfsClient.
    * @return AbfsCounters instance.
    */
-  public AbfsCounters getAbfsCounters() {
+  protected AbfsCounters getAbfsCounters() {
     return abfsCounters;
   }
