+++ hadoop-tools/hadoop-azure/src/test/java/org/apache/hadoop/fs/azurebfs/services/TestAbfsRestOperation.java
@@ -1,12 +1,11 @@
-package org.apache.hadoop.fs.azurebfs;
+package org.apache.hadoop.fs.azurebfs.services;
 
 import org.junit.Test;
 import static org.apache.hadoop.fs.azurebfs.constants.AbfsHttpConstants.HTTP_METHOD_DELETE;
 import static org.apache.hadoop.fs.azurebfs.services.AbfsRestOperationType.DeletePath;
-import org.apache.hadoop.fs.azurebfs.services.AbfsClient;
-import org.apache.hadoop.fs.azurebfs.services.TestAbfsClient;
 import java.lang.reflect.Method;
-import org.apache.hadoop.fs.azurebfs.services.AbfsRestOperation;
+import org.apache.hadoop.fs.azurebfs.AzureBlobFileSystem;
+import org.apache.hadoop.fs.azurebfs.AbstractAbfsIntegrationTest;
 import java.util.ArrayList;
 import java.util.Arrays;
 import org.junit.Assert;
@@ -21,12 +20,12 @@ public class TestAbfsRestOperation extends
   @Test
   public void testDriverRetryMetrics() throws Exception {
     final AzureBlobFileSystem fs = getFileSystem();
-    AbfsClient client = fs.getAbfsClient();
+    AbfsClient testClient = super.getAbfsClient(super.getAbfsStore(getFileSystem()));
 
     // Mock instance of AbfsRestOperation
     AbfsRestOperation op = TestAbfsClient.getRestOp(
-        DeletePath, client, HTTP_METHOD_DELETE,
-       TestAbfsClient.getTestUrl(client, "/NonExistingPath"), TestAbfsClient.getTestRequestHeaders(client));
+        DeletePath, testClient, HTTP_METHOD_DELETE,
+       TestAbfsClient.getTestUrl(testClient, "/NonExistingPath"), TestAbfsClient.getTestRequestHeaders(testClient));
 
     ArrayList<Integer> retryCounts = new ArrayList<>(Arrays.asList(35, 28, 31, 45, 10, 2, 9));
     int statusCode = HttpURLConnection.HTTP_UNAVAILABLE;
@@ -35,7 +34,7 @@ public class TestAbfsRestOperation extends
     for(int retryCount: retryCounts) {
       getMetrics.invoke(op, retryCount, statusCode);
     }
-    Assert.assertEquals(client.getAbfsCounters().getAbfsDriverMetrics().getNumberOfRequestsFailed().toString(), "3");
+    Assert.assertEquals(testClient.getAbfsCounters().getAbfsDriverMetrics().getNumberOfRequestsFailed().toString(), "3");
     fs.close();
   }
 }
